corrected get_batch()
Hello,
just stumbled across a little Mistake:
the get_batch function in Memory class always gives back a single sample, not an Batch.
the Line:
"s,a,r,gae_r,s_,d = [],[],[],[],[],[]"
was inside the for loop, so every iteration of the for loop redefined all this variables, and then appended one single Element, just for the next Iteration to redefine the Variables again. This way you only got the last sample back.